Move toast out to debugToast Abstract form context out of bank form
